hardware/mordomo - (orange-pi-5-16gb)/ecossistemas/mordomo/containers/stt/speaker-id-diarization/requirements.txt
# Core Dependencies
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0

# Speaker Recognition
resemblyzer==0.1.1.dev0
numpy==1.24.3
scipy==1.11.3

# gRPC
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.0

# NATS
nats-py==2.6.0
asyncio-nats-client==0.11.5

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Monitoring
prometheus-client==0.19.0

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
watchdog==3.0.0

# Logging
structlog==23.2.0
